add user to db when new login
