+++ functions/src/index.ts
@@ -1,33 +1,16 @@
 import * as functions from 'firebase-functions'
-import { createTransport } from 'nodemailer'
+import * as admin from 'firebase-admin'
 
-const config = functions.config().email
+admin.initializeApp()
+const db = admin.firestore()
 
-const transporter = createTransport(config)
 
-interface Message {
-    to: string
-    subject: string
-    body: string
-}
-
-async function sendMail({ to, subject, body }: Message) {
-    await transporter.sendMail({
-        from: {
-            name: 'DonateABLE',
-            address: config.auth.user,
-        },
-        to: to,
-        subject: subject,
-        html: body,
-    })
-}
-
-export const testMail = functions.https.onRequest(async (request, response) => {
-    await sendMail({
-        to: request.query.to,
-        subject: "new message",
-        body: "a new message body",
-    })
-    response.send('')
-});
+export const newUser = functions.auth
+    .user()
+    .onCreate(async user => {
+        await db.collection('users')
+            .doc(user.uid)
+            .set({
+                email: user.email,
+            })
+    });
