+++ src/components/header/style.scss
@@ -46,6 +46,11 @@
                 clip-path: circle(50px at center);
             }
         }
+
+        img {
+            width: 100%;
+            height: 100%;
+        }
     }
 
     .info {
